Rename package to pytest_html.
This is to avoid a conflict with the `html' package which is in the Python3
stdlib.
