+++ pytest_html/resources/main.js
