+++ pytest_html/resources/style.css
