+++ setup.py
@@ -7,9 +7,9 @@ setup(name='pytest-html',
       author='Dave Hunt',
       author_email='dhunt@mozilla.com',
       url='https://github.com/davehunt/pytest-html',
-      packages=['html'],
-      package_data={'html': ['resources/*']},
-      entry_points={'pytest11': ['html = html.plugin']},
+      packages=['pytest_html'],
+      package_data={'pytest_html': ['resources/*']},
+      entry_points={'pytest11': ['html = pytest_html.plugin']},
       install_requires=['pytest>=2.3'],
       license='Mozilla Public License 2.0 (MPL 2.0)',
       keywords='py.test pytest html report',
